+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Your contribution here.
 * [#2115](https://github.com/ruby-grape/grape/pull/2115): Fix declared_params regression with multiple allowed types - [@stanhu](https://github.com/stanhu).
+* [#2123](https://github.com/ruby-grape/grape/pull/2123): Fix 2.7 deprecation warning in middleware/stack - [@Legogris](https://github.com/Legogris).
 
 ### 1.5.0 (2020/10/05)